requirements.txt
# ZealX System Requirements
numpy>=1.19.0
faiss-cpu>=1.7.0
psutil>=5.9.0

# Backend Requirements
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
asyncio>=3.4.3
python-multipart>=0.0.6
starlette>=0.27.0

# Database
psycopg2-binary>=2.9.5
pgvector>=0.2.0
asyncpg>=0.27.0
SQLAlchemy>=2.0.0
alembic>=1.10.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Production Server
gunicorn>=21.2.0
uvloop>=0.17.0
httptools>=0.5.0

# Caching and Rate Limiting
redis>=4.5.0
fastapi-limiter>=0.1.5

# Monitoring and Logging
prometheus-fastapi-instrumentator>=5.9.0
python-json-logger>=2.0.0
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-exporter-otlp>=1.15.0
sentry-sdk>=1.14.0

# Security
pyjwt>=2.6.0
cryptography>=39.0.0
python-multipart>=0.0.6

# Cloud Provider SDKs
boto3>=1.26.0  # AWS
google-cloud-storage>=2.7.0  # GCP
azure-storage-blob>=12.14.0  # Azure

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.1.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0